Refactor crud package
